Update 8
Began work on post exercise controller.  Added find user method that takes the post input user id and searches the db for user by id.  After this is done child subdocument fields are updated with the post input fields and added to the subdocument children array.  Ran into an issue with mongoose not accepting the array.push method.  Will work on resolving that next before continuing.  Updated exerciseSchema to not include a uid field as mongoose auto injects a unique id field.