Update 9
Updated schema validation for dates to correctly fill in the date if a user leave the date field empty.  Updated controller function to check if the date value is an empty string and if so to set it to undefined in which case the validator will use the default.  Updated controller and schema to make sure user values are unique and display an error message if a user already exists.