Update 16
Updated call backend api function to properly update the app state with the user data from the backend api.  The render function now checks if the userdata has been fetched.  If so it calls the new userlist component else it displays a loading message.  Added userlist component that maps all the users into an unordered list containing the username and user id fields.