Update 18
Updated userform component state to include user data and a flag for if data was returned.  Updated fetch method within the component  to stringify the state data and send it as a post request.  The state is then updated with the return data from the post and lastly if the data is retrned the component shows the new user name and user id.  Began work creating an exercise form component to handle making post requests and displaying return form data.