Update 20
Removed the error logging middleware as the errors are tracked in each component now and were causing issues with header submission that was disconnecting the server.  Added exercise log fields and error message handling within exercise log form.  Added activate logs component that displays a button and if the button is clicked a form is displayed with fields to search by userid for an exercise log.  Began work on logform.  Removed req.params from the log controller as the new request will be in the form of a body instead of url params.