Update 20
Fixed issue with LogForm fetch call returning json.parse error.  The route was expecting a get request not a post.  Updated the route with the correct type.  Added variables in the handlechange function to break the event into name/value/event so each input can be added to state separately.  Created LogResults component to handle displaying the exercise log results.  The components take the userdata as props and map through the array of exercise logs which contains logs as objects.  The values are then extracted into jsx list items with labels.  Next step is to do more form validation and error handling within each component.