Update 23
Fixed bug in exercise log controller that was returning json without exerciseLogs key and causing in a typeerror when trying to return a log based on specific date range.  Added loading component and logic to each component for when data has not been returned yet to display the loading component.