Update 28
Changed get_users route to now be a post function to allow skip and result data to be passed to the controller from the front end.  Added skip and limit logic to the getusers controller.  Modified the user.find function to return total results and limit/skip in the find query.  Need to figure out how to return the total results instead of the limited results to accurately update the skip and limit as a function of total results.  This will allow the user  to display continuous results