Update 36
Added fonts and colors.  Added responsiveness to the css.  Made it so the grid is adjusted to one vertical column with 4 equal rows when the screen size gets below 768px.  Began work adding button colors, fonts and adjusting sizes for buttons and container elements.  Adjusted the footer to relative as it wasn't properly hanging to the bottom when screen size changes due to dynamic content.  Fixed bug with previous results that was happening because prevResults flag wasn't being reset in the get-user controller.  Added display form button and functionality for the create username form.