Update 37
Adjusted spacing for form elements and inputs.  Adjusted styling for li and ul elements to remove symbols and align left.  Began testing borders.  Adjusted spacing and margins.  Found an issue with footer not correctly being placed at bottom of the page and being over-written in mobil when result data is displayed